Rename pinniped-server -> pinniped-concierge
Do we like this? We don't know yet.
